# tests/docs/openapi.json
{
    "openapi": "3.0.2",
    "info": {
        "title": "Toybank",
        "version": "0.1.0"
    },
    "paths": {
        "/customers": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get All Customers",
                "operationId": "get_all_customers_customers_get",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers"
                ],
                "summary": "Create Customer",
                "operationId": "create_customer_customers_post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{customer_id}": {
            "get": {
                "tags": [
                    "Customers"
                ],
                "summary": "Get One Customer",
                "operationId": "get_one_customer_customers__customer_id__get",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Customer Id",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{customer_id}/accounts": {
            "get": {
                "tags": [
                    "Customers",
                    "Accounts"
                ],
                "summary": "Get Customer Accounts",
                "operationId": "get_customer_accounts_customers__customer_id__accounts_get",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Customer Id",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Customers",
                    "Accounts"
                ],
                "summary": "Create Customer Account",
                "operationId": "create_customer_account_customers__customer_id__accounts_post",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Customer Id",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerAccountCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Account"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/customers/{customer_id}/accounts/{account_id}/transfers": {
            "post": {
                "tags": [
                    "Customers",
                    "Accounts",
                    "Transfers"
                ],
                "summary": "Create Transfer",
                "operationId": "create_transfer_customers__customer_id__accounts__account_id__transfers_post",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Customer Id",
                            "type": "string"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Account Id",
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/CustomerAccountTransferCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Transfer"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Customers",
                    "Accounts",
                    "Transfers"
                ],
                "summary": "Get Account Transfers",
                "operationId": "get_account_transfers_customers__customer_id__accounts__account_id__transfers_get",
                "parameters": [
                    {
                        "name": "customer_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Customer Id",
                            "type": "string"
                        }
                    },
                    {
                        "name": "account_id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "title": "Account Id",
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        }
    },
    "components": {
        "schemas": {
            "Account": {
                "title": "Account",
                "required": [
                    "alias",
                    "customer_id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created At",
                        "type": "string",
                        "format": "date-time"
                    },
                    "alias": {
                        "title": "Alias",
                        "type": "string"
                    },
                    "balance": {
                        "title": "Balance",
                        "type": "integer",
                        "default": 0
                    },
                    "customer_id": {
                        "title": "Customer Id",
                        "type": "string"
                    }
                }
            },
            "CustomerAccountCreate": {
                "title": "CustomerAccountCreate",
                "required": [
                    "alias"
                ],
                "type": "object",
                "properties": {
                    "alias": {
                        "title": "Alias",
                        "type": "string"
                    },
                    "balance": {
                        "title": "Balance",
                        "type": "integer",
                        "default": 0
                    }
                }
            },
            "CustomerAccountTransferCreate": {
                "title": "CustomerAccountTransferCreate",
                "required": [
                    "recipient_account_id",
                    "amount"
                ],
                "type": "object",
                "properties": {
                    "recipient_account_id": {
                        "title": "Recipient Account Id",
                        "type": "string"
                    },
                    "amount": {
                        "title": "Amount",
                        "exclusiveMinimum": 0.0,
                        "type": "integer"
                    }
                }
            },
            "CustomerCreate": {
                "title": "CustomerCreate",
                "required": [
                    "email"
                ],
                "type": "object",
                "properties": {
                    "email": {
                        "title": "Email",
                        "type": "string"
                    }
                }
            },
            "HTTPValidationError": {
                "title": "HTTPValidationError",
                "type": "object",
                "properties": {
                    "detail": {
                        "title": "Detail",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        }
                    }
                }
            },
            "Transfer": {
                "title": "Transfer",
                "required": [
                    "recipient_account_id",
                    "amount",
                    "sender_account_id"
                ],
                "type": "object",
                "properties": {
                    "id": {
                        "title": "Id",
                        "type": "string"
                    },
                    "created_at": {
                        "title": "Created At",
                        "type": "string",
                        "format": "date-time"
                    },
                    "recipient_account_id": {
                        "title": "Recipient Account Id",
                        "type": "string"
                    },
                    "amount": {
                        "title": "Amount",
                        "exclusiveMinimum": 0.0,
                        "type": "integer"
                    },
                    "sender_account_id": {
                        "title": "Sender Account Id",
                        "type": "string"
                    }
                }
            },
            "ValidationError": {
                "title": "ValidationError",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "type": "object",
                "properties": {
                    "loc": {
                        "title": "Location",
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    },
                    "msg": {
                        "title": "Message",
                        "type": "string"
                    },
                    "type": {
                        "title": "Error Type",
                        "type": "string"
                    }
                }
            }
        }
    }
}